[Button] Rename TpeDefault to TpeButton
For the same reason why we have ColorWhite instead of ColorDefault: The name of something should reflect its meaning, not its usage.

The fact that we use it as the default value or not should not affect its value at all.

"The default value of Tpe is TpeButton" sounds much more reasonable than "The default value of Tpe is TpeDefault"