[Cpn] Tooltip: avoid overflow layout in initial render
In initial render of Tooltip's content, the correct position is unknown, so the content is placed at the end of the Portal's div, or in simpler words: outside of body tag.

This created a scroll bar in body.

This PR add (top: 0, left: 0) position to the initial render to avoid that (Tooltip's content will be rendered to top left, hidden, in initial render). Popper will then apply the correct style (position) from 2nd render