[Platform] Popover should support isFullWidth
For 95% of the time, `width: max-content` is what we will need.

This is because `Popover` wraps its target inside a `div`. Without `width: max-content`, the width that Popper uses to calculate will be 100%, although the size of the target usually is not (for example, a Button).

However, the down side of `width: max-content` parent is that it prevents children from having 100% width. Therefore, we still need an option to have Popover with 100% width, so its children will also be