[Platform] Component: Rename TextInput to TextBox
We are following web naming, so variation of HTML's `input` tag should be named as the `type` it represents, for example:
- Checkbox
- Radio
- Text

We choose to name this TextBox instead of Text because Text is simply too generic in most contexts.

We avoid naming it "TextField" since we want "Field" to represent something that also has label and description and error message.

We avoid using "Input" in naming to keep it consistent with other input components (like Checkbox and Radio). All input components are grouped under `input` package, but not having `Input` in their name.

That being said, it worth to note that `Button` and `Dropdown` are not grouped under `input` because they are not HTML's `input` tag, but `button` and `select` tag, respectively.