[Platform] Dropdown: Simplify RenderOption
Initially `renderOption` allows engineer to completely define the render result of each option. In practice, this turns out to be bad because the engineer then needs to do too many things.

In practice, they only need to have something different than `renderValue`, while still having all the out-of-the-box benefits like interactions control by downshift.

This commit does just that: `renderOption` is simply `renderValue` that would be used to render option list. If None is provided, it just fallbacks to renderValue