fix(platform): TextBox: Improve value controller
In #19325 I updated TextBox to always control its value. The idea is correct (so far) but the implementation has an issue that the final "input" tag will have both "defaultValue" (set by TextMask) and "value" (set by us).

This PR is to fix it with the correct way, that is the rename the prop from "defaultValue" to "value" after received from TextMask and before applying it.

This is also the author's recommendation. Please see detail in code's comment