refactor(platform): Button: Rename Size to Height
The name "Size" implies it controls both Height and Width, which is a wrong thing. The current "size" prop [only controls the Height](https://anduin.design/button#size) and some related styles, while the "width" is controled in [another prop](https://anduin.design/button#full-width) already.

This PR is to eliminate this misunderstanding by renaming Size to Height, which is the thing that it actually defines