feat(platform): Modal: Support custom Height
Currently there are 2 choices regarding Modal's height:
- Use Modal's content, or
- 100% of screen

This is not really nice, because consumers cannot have anything like "a percentage of screen height".

This PR improves the implementation of Modal's Height to support just that, with a `case class Custom(percent: Int)`. `Full` is now just a special case `Custom(100)`.

This allows consumers like CommentButton to, for example, having a 80% screen Modal