fix(platform): Style: Remove Disabled
Disabled styling should be control via JS, not CSS. Why? Because we might want to target more than just the host component.

For example, in Checkbox, the label is the parent of the input tag, but the input tag is the one who will trigger ":disabled". It's therefore impossible to gray out the label from the input