fix(platform): Radio: No need for name and value
In normal HTML, or in React's uncontrolled situation, the "checked" state of a radio is controlled by the browser (e.g. when one is checked then others in the same group will be unchecked automatically).

However, since our Radio component is stateless (as it should be), the consumers already responsible to control these "checked" state. The "name" and "value" properties, therefore, are useless in practice.

I should have realized this sooner, since all Radio's usages are simply skipped the "value" param in their "onChange" definition. Something like this:

```scala
Radio(
  onChange = _ => doThing,
  /* ... */
)()
```

Moreover, these props are not only useless but also open the possibility to doing things wrong without seeing the consequence. For example, you could set the "name" prop of 2 radios differently, but still "connect" them properly via "isChecked" and "onChange".

In fact, "isChecked" and "onChange" should be enough for all practical cases