fix(platform): Modal should have focus style
It's able for Modal to have user's focus since a long time (so features like "Esc to close" is possible). Until now, it uses the default focus style, which is a blue ring in Mac but an orange ring in Ubuntu.

This PR makes It to use use the same styles with other components. Note that since we have "focus-visible", the focus style won't be visible until users actually need it (e.g. using keyboard)