fix(platform): Dropdown: fix lost search value when using keyboard
This is a very subtle bug that I randomly saw while working with Downshift :joy:

Basically we have a small logic in Dropdown to clear the input value (in search box) when users open the menu. This is because Downshift automatically set the input value to the currently selected option, and without clearing it others options won't be shown because of filtering.

The problem is our current logic is not good enough. It only check if (1) the menu is going to be opened and (2) the user is not typing.

This results in a false negative where users typed something, then use keyboard navigation to select from the filtered list. In this case, both checks are ok and the search value will be cleared, although that is not what we want.

This PR improves the check by simplify it: the search value should only be cleared when the menu is opened or closed, and that's all. This works much better: https://anduin.design/dropdown#search