fix(platform): Downshift: Move "merge" to Downshift
Merging downshift's state changes is a little bit more complicated
in ScalaJS, mostly because it doesn't support ES6's `Object.assign`.
This is why we have this convenient util.

However, it's better to have it in Downshift instead of Dropdown since
it's quite generic and Suggest should also be able to use it